+++ packages/styleguide/.storybook/addons/system/components/Elements.tsx
@@ -1,14 +1,16 @@
 import styled from '@emotion/styled';
-import { themed, variant } from '@codecademy/gamut-styles/src';
+import { themed, system } from '@codecademy/gamut-styles/src';
+import { StyleProps } from '@codecademy/variance';
 
 export const PropGroupTooltip = styled.ul`
   display: none;
   position: absolute;
   bottom: calc(100% + 10px);
   left: 50%;
+  min-width: 80px;
   transform: translate(-50%, 0);
   padding: ${themed('spacing.8')} ${themed('spacing.12')};
-  border: 1px solid ${themed('colors.navy')};
+  border: ${themed('borders.1')};
   background-color: ${themed('colors.white')};
   list-style: none;
   margin: 0;
@@ -17,36 +19,37 @@ export const PropGroupTooltip = styled.ul`
 
 export const PropItem = styled.li`
   margin: ${themed('spacing.8')} 0;
-  font-size: 11px;
-  font-family: ${themed('fontFamily.accent')};
+  font-size: ${themed('fontSize.14')};
   text-align: center;
   text-transform: none;
   color: ${themed('colors.gray-700')};
 `;
 
-const tagVariants = variant({
-  normal: {
-    textColor: 'navy',
-    borderColor: 'navy',
-  },
-  selected: {
-    textColor: 'white',
-    backgroundColor: 'navy',
-    borderColor: 'navy',
+const tagVariants = system.variant({
+  variants: {
+    normal: {
+      textColor: 'navy',
+      borderColor: 'navy',
+    },
+    selected: {
+      textColor: 'white',
+      bg: 'navy',
+      borderColor: 'navy',
+    },
   },
 });
 
-export const PropGroupTag = styled.span<Parameters<typeof tagVariants>[0]>`
+export const PropGroupTag = styled.span<StyleProps<typeof tagVariants>>`
   ${tagVariants}
   user-select: none;
   position: relative;
   display: inline-block;
   padding: ${themed('spacing.4')} ${themed('spacing.8')};
-  font-size: 11px;
-  font-family: ${themed('fontFamily.accent')};
+  font-size: ${themed('fontSize.14')};
+  font-family: ${themed('fontFamily.base')};
   border-radius: 4px;
   cursor: help;
-  border: 1px solid;
+  border: ${themed('borders.1')};
 
   &:hover {
     > * {