+++ packages/gamut-illustrations/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.13.3](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-illustrations@0.13.2...@codecademy/gamut-illustrations@0.13.3) (2021-05-04)
+
+**Note:** Version bump only for package @codecademy/gamut-illustrations
+
+
+
+
+
+### [0.13.2](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-illustrations@0.13.1...@codecademy/gamut-illustrations@0.13.2) (2021-05-04)
+
+**Note:** Version bump only for package @codecademy/gamut-illustrations
+
+
+
+
+
 ### [0.13.1](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-illustrations@0.13.0...@codecademy/gamut-illustrations@0.13.1) (2021-05-03)
 
 **Note:** Version bump only for package @codecademy/gamut-illustrations